docs(readme): add minimum software req to README
Based on reviewer feedback, many of the installation challenges stemmed
from the fact that our instructions were written for the latest LTS
version of Ubuntu (18.04), and not an earlier version. It would be
difficult to write instructions for each previous version, so instead we
specify the expected minimum software versions for which these
instructions have been tested for.